Increasing uptime counter to 64 bits

+++ src/Components/Statistics/Statistics.h
@@ -6,7 +6,7 @@ struct Statistics
     uint32_t NumModbusWriteErrors = 0;
     uint32_t NumWifiDisconnects = 0;
     uint32_t NumMqttDisconnects = 0;
-    uint32_t UptimeS = 0;
+    uint64_t UptimeS = 0;
 };
 
 // Global statistics instance
